Внес поправки от наставника в базовой верстке react
