+++ src/components/CostAnalysis/Header.jsx
@@ -1,0 +1,9 @@
+import styles from "./CostAnalysis.module.css";
+
+const Header = () => (
+  <header className={styles.pageHeader} role="banner">
+    <h1 className={styles.pageTitle}>Анализ расходов</h1>
+  </header>
+);
+
+export default Header;
