+++ src/components/CostAnalysis/constants.js
@@ -1,0 +1,16 @@
+export const MONTHS = [
+  "Январь",
+  "Февраль",
+  "Март",
+  "Апрель",
+  "Май",
+  "Июнь",
+  "Июль",
+  "Август",
+  "Сентябрь",
+  "Октябрь",
+  "Ноябрь",
+  "Декабрь",
+];
+
+export const WEEKDAYS = ["пн", "вт", "ср", "чт", "пт", "сб", "вс"];
